github-automation/requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
google-cloud-storage==2.14.0
